fix: Fix issues with uploading code to ESP8266
Corrected issues associated with not being able to upload unit test
code to ESP8266.
* Found bugs in latest PlatformIO release of ESP8266
Arduino (2.2.0).  Switched code to prerelease (actual ESP8266 framework
is up to version 2.6.0.
* Changed test main to loop execute the tests instead of one shot.
Allows for easier debugging of tests.  When ran in a one shot, the UART
output was not being flushed unless you reset the board.
* Fixed src_filter issue where unit test code wasn't being included
in the build due to a merge error.